pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lore-ontology"
version = "0.2.0"
description = "Human-readable ontology format for the AI age"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
lore = "lore.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
